UAT_Smoke_tests
