+++ apps/playwright/tests/smoke/platform-admin-smoke.spec.ts
@@ -1,5 +1,6 @@
 import { test, expect } from "~/fixtures";
 import { loginBySub } from "~/utils/auth";
+import AUTH from "~/constants/auth";
 
 const sub =
   process.env.PLAYWRIGHT_PLATFORM_ADMIN_SUB ??
@@ -19,5 +20,22 @@ test.describe("Platform admin smoke", { tag: "@uat" }, () => {
         name: /welcome back/i,
       }),
     ).toBeVisible();
+  });
+
+  // Moved from auth > admin.spec.ts file
+  test("Can access restricted paths", async ({ appPage }) => {
+    await Promise.all(
+      AUTH.RESTRICTED_PATHS.ADMIN.map(async (restrictedPath) => {
+        const context = appPage.page.context();
+        const page = await context.newPage();
+        await page.goto(restrictedPath);
+        await page.waitForURL(restrictedPath);
+        await expect(
+          page.getByRole("heading", {
+            name: "Sorry, you are not authorized to view this page.",
+          }),
+        ).toBeHidden();
+      }),
+    );
   });
 });
